Fixed MySQL start configuration

+++ myproject-core/src/main/java/net/andresbustamante/myproject/core/dao/ActorDao.java
@@ -4,5 +4,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import net.andresbustamante.myproject.core.entities.Actor;
 
-public interface ActorDao extends JpaRepository<Actor, Integer> {
+public interface ActorDao extends JpaRepository<Actor, Short> {
 }
