+++ myproject-core/src/main/java/net/andresbustamante/myproject/core/dao/StaffDao.java
@@ -6,7 +6,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import net.andresbustamante.myproject.core.entities.Staff;
 
-public interface StaffDao extends JpaRepository<Staff, Integer> {
+public interface StaffDao extends JpaRepository<Staff, Short> {
 
     List<Staff> findAllByActiveTrue();
 }
